+++ src/Kit/DataEntry/RichText/EditorContentCharacterCount/types.d.ts
@@ -1,0 +1,8 @@
+import {type Editor} from '@tiptap/core';
+import {KitRichTextStatus} from '../types';
+
+export type IKitEditorContentCharacterCount = {
+    editor: Editor;
+    maxLength?: number;
+    status?: KitRichTextStatus;
+};