+++ src/Kit/DataEntry/RichText/MenuBar/FontStylesButtons/types.d.ts
@@ -1,0 +1,5 @@
+import {type Editor} from '@tiptap/core';
+
+export type IKitFontStylesButtons = {
+    editor: Editor;
+};