+++ src/translation/fr-FR/components/DataEntry/RichText/index.ts
@@ -1,0 +1,29 @@
+import {IKitRichTextLocale} from '@translation/types/components/DataEntry/RichText';
+
+export const KitRichTextLocale: IKitRichTextLocale = {
+    undo: 'Annuler',
+    redo: 'Rétablir',
+    bold: 'Gras',
+    italic: 'Italique',
+    underline: 'Souligner',
+    alignLeft: 'Aligner à gauche',
+    alignCenter: 'Aligner au centre',
+    alignRight: 'Aligner à droite',
+    justify: 'Justifier',
+    insertLink: 'Insérer un lien',
+    bulletList: 'Liste à puces',
+    orderedList: 'Liste ordonnée',
+    indent: 'Augmenter le retrait',
+    outdent: 'Réduire le retrait',
+    heading1: 'En-tête 1',
+    heading2: 'En-tête 2',
+    heading3: 'En-tête 3',
+    heading4: 'En-tête 4',
+    normal: 'Normal',
+    textToDisplay: 'Texte à afficher',
+    url: 'URL',
+    linkToDocumentOrWebPage: 'Lien vers un document ou une page Web',
+    invalidUrl: 'URL invalide',
+    textToDisplayCannotBeEmpty: 'Le texte à afficher ne peut pas être vide',
+    urlCannotBeEmpty: "L'URL ne peut pas être vide"
+};