+++ src/translation/types/index.d.ts
@@ -6,6 +6,7 @@ import {IKitEmptyLocale} from './components/DataDisplay/Empty';
 import {IKitErrorLocale} from './components/DataDisplay/Error';
 import {IKitGeneralLocale} from './general';
 import {IKitAlertLocale} from './components/Feedback/Alert';
+import {IKitRichTextLocale} from './components/DataEntry/RichText';
 
 type LanguageSupported = 'frFR' | 'enUS';
 
@@ -18,4 +19,5 @@ export interface IKitLocale extends Locale {
     Error?: IKitErrorLocale;
     Image?: IKitImageLocale;
     ItemList?: IKitItemListLocale;
+    RichText?: IKitRichTextLocale;
 }