+++ stories/02-Design System/04-DataEntry/RichText/examples/disabled.tsx
@@ -1,0 +1,10 @@
+import {KitRichText} from '@kit/DataEntry';
+import {KitSpace} from '@kit/Layout';
+
+const App = () => (
+    <KitSpace>
+        <KitRichText placeholder="PlaceHolder..." disabled />
+    </KitSpace>
+);
+
+export default App;