+++ stories/02-Design System/04-DataEntry/RichText/test-components/index.tsx
@@ -1,0 +1,5 @@
+export {BasicRichTextTest} from './BasicRichTextTest';
+export {CharacterCountRichTextTest} from './CharacterCountRichTextTest';
+export {DisabledRichTextTest} from './DisabledRichTextTest';
+export {LabelAndHelperRichTextTest} from './LabelAndHelperRichTextTest';
+export {ReadOnlyRichTextTest} from './ReadOnlyRichTextTest';